requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
regex==2024.5.15
python-multipart==0.0.9
python-dotenv==1.0.1
# Optional OCR (app runs without these)
pillow==10.3.0
pytesseract==0.3.10